Updated CraftMail.lua and CraftMailServer.lua
The left part in the menu is now integrated (no touch screen for the moment though). The client is now more in sync with the server (Data exchanges were simplified and are now with the same generic shape). The top-screen bar has a title and the name of the account ID.